Split ReloadIntervalSelect and TrialTable from StudyDetail.tsx
